move read and write json functions to separate file
