+++ server/api.go
@@ -147,14 +147,3 @@ func ToHttpHandlerFunc(f APIFunc) http.HandlerFunc {
 		}
 	}
 }
-
-func WriteJSON(w http.ResponseWriter, status int, v any) error {
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(status)
-
-	return json.NewEncoder(w).Encode(v)
-}
-
-func ReadJSON(r *http.Request, v any) error {
-	return json.NewDecoder(r.Body).Decode(v)
-}
