+++ server/json.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+func WriteJSON(w http.ResponseWriter, status int, v any) error {
+	w.Header().Add("Content-Type", "application/json")
+	w.WriteHeader(status)
+
+	return json.NewEncoder(w).Encode(v)
+}
+
+func ReadJSON(r *http.Request, v any) error {
+	return json.NewDecoder(r.Body).Decode(v)
+}
